refactor: main.swift의 코드를 ConsoleApp 클래스에 분리
